- Reimplementando protocolo e tratamento da recepção dos dados pela serial .....- PC aguarda CONNECT, responde OK e espera por READY, arduino só inicia medições após receber START .....- Criado queue para servir de buffer na recepção dos dados, garantindo a leitura correta da porta serial .....- Criada classe estendida do SerialPort para melhor modularizar a implementação da comunicação serial - Debug dos milisegundos adicionado à janela de teste (sincronia do gráfico toma o primeiro como referência) - carro agora envia o "millis()" dos dados capturados para maior precisão na exibição dos gráficos

+++ src/TeleBajaUEA-client/TeleBajaUEA/RaceDataStructs/FileSensorsData.cs
@@ -9,7 +9,7 @@ namespace TeleBajaUEA.RaceDataStructs
     [Serializable()]
     public struct FileSensorsData
     {
-        public float xValue;
+        public uint xValue;
         public float speed;
         public float rpm;
         public bool breakState;
